修改int为C++的size_t

+++ C++/h/cdlist.h
@@ -39,7 +39,7 @@ public:
 	}
 
 private:
-	int size;
+	size_t size;
 	DListElmt *head;
 	DListElmt *tail;
 };
